Did a bit more. it's 4:21 rn
